pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Agentic-TAF"
version = "1.0.0"
description = "Agentic Test Automation Framework"
readme = "README.md"
license = "LGPL-3.0-or-later"
requires-python = ">=3.12"
authors = [
    {name = "Wesley Peng", email = "wesley.peng@live.com"},
]
keywords = ["test-automation", "framework", "playwright", "api-testing", "llm-testing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "PyYAML>=6.0",
    "paramiko>=3.0",
    "requests>=2.31",
    "selenium>=4.20",
]

[project.optional-dependencies]
appium = ["Appium-Python-Client>=4.0"]
playwright = ["playwright>=1.40"]
httpx = ["httpx>=0.27"]
websocket = ["websockets>=12.0"]
llm = ["langchain-anthropic>=0.3"]
all = [
    "Appium-Python-Client>=4.0",
    "playwright>=1.40",
    "httpx>=0.27",
    "websockets>=12.0",
    "langchain-anthropic>=0.3",
]
dev = [
    "pytest>=8.0",
    "flake8>=7.0",
    "mypy>=1.10",
    "types-PyYAML>=6.0",
    "types-paramiko>=3.0",
    "types-requests>=2.31",
    "behave>=1.2",
    "robotframework>=7.0",
    "allure-behave>=2.13",
]

[tool.setuptools.packages.find]
where = ["src/main/python"]
exclude = ["ut*", "bpt*"]

[tool.setuptools.package-data]
taf = ["foundation/conf/config.yml"]

[tool.pytest.ini_options]
testpaths = ["src/test/python"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "e2e: end-to-end tests (require live cluster)",
    "chaos: chaos engineering tests",
    "load: load/performance tests",
    "ai: AI-specific tests (require LLM access)",
]

[tool.flake8]
max-line-length = 120
exclude = [".git", "__pycache__", "build", "dist", "*.egg-info"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
check_untyped_defs = true
